% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmp-package.R
\docType{package}
\name{cmp-package}
\alias{cmp}
\alias{cmp-package}
\title{Conway-Maxwell Poisson Distribution}
\description{
Provides routines for computing the density of the Conway-Maxwell Poisson
distribution and fitting parameters to data.
}
\details{
\tabular{ll}{ Package: \tab cmp\ cr Type: \tab Package\cr Version:
\tab 0.2\cr Date: \tab 2008-04-21\cr License: \tab BSD\cr }
}
\author{
Noam Ross
}
\references{
Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and
Boatwright, P., \dQuote{A useful distribution for fitting discrete data:
Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc.,
v54, pp. 127-142, 2005.
}
\keyword{models}
\keyword{package}
\keyword{regression}

