% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmp_log_mean}
\alias{cmp_log_mean}
\title{Computes Expectation of a Function of a CMP Random Variable}
\usage{
cmp_log_mean(lambda, nu, log_error = 1e-06, maxit = 1000000L,
  z = NA_real_, log_error_z = 1e-06, maxit_z = 10000L, parallel = FALSE)
}
\arguments{
\item{lambda}{value of lambda parameter}

\item{nu}{value of nu parameter}

\item{f}{function taking as a single argument the value of x}

\item{log.error}{precision in the log of the expectation}
}
\value{
The expectation as a real number.
}
\description{
Computes an expectation of a function of a CMP random variable.
}
\details{
Computes the expectation \eqn{E[f(X)]}{E[f(X)]} where X is a CMP
random variable.
}
\author{
Jeffrey Dunn
}
\references{
Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and
Boatwright, P., \dQuote{A useful distribution for fitting discrete data:
Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc.,
v54, pp. 127-142, 2005.
}
\seealso{
\code{\link{cmp_mean}}, \code{\link{cmp_var}},
\code{\link{cmp_fit}}
}
\keyword{models}

