% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/com.R
\name{com.log.density}
\alias{com.log.density}
\title{Computes the Log PMF of the COM-Poisson Distribution}
\usage{
com.log.density(x, lambda, nu, log.z = NULL)
}
\arguments{
\item{x}{level to evaulate the log PMF at}

\item{lambda}{value of the lambda parameter}

\item{nu}{value of the nu parameter}

\item{log.z}{log of the normalizing constant, computed if not specified}
}
\value{
The log probability that a random COM-Poisson variable X takes value
x.
}
\description{
Computes the log probability mass function of the COM-Poisson distribution
for given values of the parameters.
}
\details{
Computes the log probability mass function of the COM-Poisson distribution
\deqn{ }{log f(x) = x * log(lambda) - log(Z) - nu * log(x!). }\deqn{ \log
f(x) = x \log \lambda - \log(Z(\lambda,\nu)) - \nu \sum_{i=1}^x x. }{log
f(x) = x * log(lambda) - log(Z) - nu * log(x!). }\deqn{ }{log f(x) = x *
log(lambda) - log(Z) - nu * log(x!). }
}
\examples{
data(insurance);
	fit = com.fit(Lemaire);
	com.log.density(0, fit$lambda, fit$nu, fit$z);
}
\author{
Jeffrey Dunn
}
\references{
Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and
Boatwright, P., \dQuote{A useful distribution for fitting discrete data:
Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc.,
v54, pp. 127-142, 2005.
}
\seealso{
\code{\link{com.loglikelihood}}, \code{\link{dcom}}
}
\keyword{models}

