% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log.space.R
\name{com.log.difference}
\alias{com.log.difference}
\alias{com.log.factorial}
\alias{com.log.sum}
\title{Operations in Log-space}
\usage{
com.log.difference(x, y)
}
\arguments{
\item{x}{first value}

\item{y}{second value}
}
\value{
The requested computation in log-space.
}
\description{
Computes the difference of two values in log-space.
}
\details{
\code{com.log.difference} computes the difference of two values in
log-space, \eqn{log( e^x - e^y )}, without significant chance of overflow or
underflow.

\code{com.log.sum} computes the sum of two values in log-space, \eqn{log(
e^x + e^y )}, without significant change of overflow or underflow.

\code{com.log.factorial} computes \eqn{log(x!)} which is equivalent to a
summation.
}
\examples{
a = exp(com.log.difference(log(100), log(20))); # a = 80
	b = exp(com.log.sum(log(100), log(20))); # b = 120
	c = exp(com.log.factorial(4)); # c = 24
}
\author{
Jeffrey Dunn
}
\keyword{manip}

