% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/com.moments.R
\name{com.mean}
\alias{com.mean}
\title{Computes Mean of the COM-Poisson Distribution}
\usage{
com.mean(lambda, nu, ...)
}
\arguments{
\item{lambda}{value of lambda parameter}

\item{nu}{value of the nu parameter}
}
\value{
The mean of the distribution.
}
\description{
Computes the mean of the COM-Poisson distribution for given values of the
parameters.
}
\details{
Uses \code{\link{com.expectation}} to compute the first moment of the
distribution.
}
\examples{
data(insurance)
  model = com.fit(Lemaire)
  com.mean(model$lambda, model$nu)
}
\author{
Jeffrey Dunn
}
\references{
Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and
Boatwright, P., \dQuote{A useful distribution for fitting discrete data:
Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc.,
v54, pp. 127-142, 2005.
}
\seealso{
\code{\link{com.expectation}}, \code{\link{com.var}}
}
\keyword{models}

