% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/com.moments.R
\name{com.var}
\alias{com.var}
\title{Computes Variance of the COM-Poisson Distribution}
\usage{
com.var(lambda, nu, ...)
}
\arguments{
\item{lambda}{value of lambda parameter}

\item{nu}{value of the nu parameter}
}
\value{
The variance of the distribution.
}
\description{
Computes the variance of the COM-Poisson distribution for given values of
the parameters.
}
\details{
Uses \code{\link{com.expectation}} to compute the second moment of the
distribution and subtracts the squared mean, computed using
\code{\link{com.mean}}.
}
\examples{
data(insurance)
  model = com.fit(Lemaire)
  com.var(model$lambda, model$nu)
}
\author{
Jeffrey Dunn
}
\references{
Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and
Boatwright, P., \dQuote{A useful distribution for fitting discrete data:
Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc.,
v54, pp. 127-142, 2005.
}
\seealso{
\code{\link{com.expectation}}, \code{\link{com.mean}}
}
\keyword{models}

