% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcmp}
\alias{dcmp}
\alias{pcmp}
\alias{qcmp}
\alias{rcmp}
\title{The Conway-Maxwell-Poisson Distribution}
\usage{
dcmp(x, lambda, nu, z = NA_real_, log_p = FALSE, log_error_z = 1e-06,
  maxit_z = 10000L, parallel = FALSE)

pcmp(q, lambda, nu, z = NA_real_, log_p = FALSE, log_error_z = 1e-06,
  maxit_z = 10000L, parallel = FALSE)

qcmp(p, lambda, nu, z = NA_real_, log_p = FALSE, log_error_z = 1e-06,
  maxit_z = 10000L, parallel = FALSE)

rcmp(n, lambda, nu, z = NA_real_, log_error_z = 1e-06, maxit_z = 10000L,
  parallel = FALSE)
}
\arguments{
\item{x}{level to evaluate the PMF at}

\item{lambda}{value of lambda parameter}

\item{nu}{value of nu parameter}

\item{z}{normalizing constant, computed if not specified}

\item{n}{number of random values to return}

\item{log.z}{natural log of z}
}
\value{
\code{dcmp} gives the probability that a random CMP variable
X takes value x.  \code{rcmp} gives a vector of \code{n} random values
sampled from the CMP distribution.
}
\description{
Probability mass function and random generation for the Conway-Maxwell-
Poisson Distribution distribution for given values of the parameters.
}
\details{
Computes the probability mass function of the CMP distribution
\deqn{ }{f(x) = (1/Z) (lambda^x)/(x!^nu).}\deqn{ f(x) =
\frac{1}{Z(\lambda,\nu)}\frac{\lambda^x}{(x!)^\nu}. }{f(x) = (1/Z)
(lambda^x)/(x!^nu).}\deqn{ }{f(x) = (1/Z) (lambda^x)/(x!^nu).}
}
\author{
Jeffrey Dunn
}
\references{
Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and
Boatwright, P., \dQuote{A useful distribution for fitting discrete data:
Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc.,
v54, pp. 127-142, 2005.
}
\seealso{
\code{\link{cmp_loglikelihood}},
}
\keyword{models}

