% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_cmp.R
\name{DIC_cmp}
\alias{DIC_cmp}
\title{DIC of the regression model}
\usage{
DIC_cmp(fit, X, y, S = 100)
}
\arguments{
\item{fit}{An object from the mcmc_cmp_mh}

\item{X}{Covariates list, each element is the design matrix for each column of y}

\item{y}{Matrix of observations}

\item{S}{Number of iterations used to calculate the DIC}
}
\value{
Approximated DIC
}
\description{
DIC of the regression model
}
