% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epl_20_21}
\alias{epl_20_21}
\title{Scores English Premier League Season 2020-2021}
\format{
A data frame with 380 rows and 4 variables:
\describe{
  \item{HG}{Goals scored by home team.}
  \item{AG}{Goals scored by away team.}
  \item{HomeTeam}{Home team.}
  \item{AwayTeam}{Away team.}
}
}
\source{
<https://www.football-data.co.uk>
}
\usage{
epl_20_21
}
\description{
A data set with the scores of the games played during season 2020-2021 in the English Premier League (EPL)
}
\keyword{datasets}
