% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_cmp.R
\name{llk_cmp}
\alias{llk_cmp}
\title{Log likelihood of the Conway-Maxwell-Poisson Distribution}
\usage{
llk_cmp(y, mu, nu, r = 1000)
}
\arguments{
\item{y}{Count value}

\item{mu}{Location parameter}

\item{nu}{Shape parameter}

\item{r}{Number of acceptances}
}
\value{
Estimation of the log likelihood of the distribution
}
\description{
This function calculates the log likelihood of the distribution as described by Benson and Friel (2021)
}
\examples{
llk_cmp(10, 5, 2)
}
