% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cmp.R
\name{log_cmp}
\alias{log_cmp}
\title{Log density of the normalized component of the Conway-Maxwell-Poisson}
\usage{
log_cmp(y, mu, nu)
}
\arguments{
\item{y}{Value}

\item{mu}{Location parameter}

\item{nu}{Shape parameter}
}
\value{
Numeric corresponding to the log of the unnormalized component of the distribution
}
\description{
Log density of the normalized component of the Conway-Maxwell-Poisson
}
