% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_coxph_surv.R
\name{predict_coxph_surv}
\alias{predict_coxph_surv}
\title{Obtain survival estimates from a \code{coxph()} model}
\usage{
predict_coxph_surv(mod, newdata, time)
}
\description{
Obtain survival estimates from a \code{coxph()} model
}
