% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cancer_death_risk.R
\name{prob_to_exp_lambda}
\alias{prob_to_exp_lambda}
\title{Convert Prob. to an Exponential lambda Parameter}
\usage{
prob_to_exp_lambda(p, time = 1)
}
\arguments{
\item{p}{numeric vector of probability estimates}

\item{time}{a single numeric time that is the time frame for the probability
in \verb{p=}}
}
\value{
numeric lambda greater than zero
}
\description{
Function takes a vector of probability estimates and a time frame,
e.g. probability of death within 10 years, and converts the probability
into an exponential distribution lambda parameter. The lambda parameter
is the \emph{rate} parameter, e.g. \code{f(x) = lambda * exp(-lambda * x)}
}
\examples{
prob_to_exp_lambda(0.5)
}
