% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_list.R
\name{prod_list}
\alias{prod_list}
\title{Product of lists between matrices}
\usage{
prod_list(X, beta)
}
\arguments{
\item{X}{Data}

\item{beta}{Parameters}
}
\value{
A list with the products element-wise
}
\description{
Product of lists between matrices
}
