% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_cmp.R
\name{DIC_cmp}
\alias{DIC_cmp}
\title{DIC of the regression model}
\usage{
DIC_cmp(fit, S = 100)
}
\arguments{
\item{fit}{An object from the mcmc_cmp_mh}

\item{S}{Number of iterations used to calculate the DIC}
}
\value{
Approximated DIC
}
\description{
This function used an approach similar to the presented by Benson & Friel (2021) to calculate the BIC. We select S a sample size of the posterior samples to speed up computation
}
