% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultRegCMP-package.R
\docType{package}
\name{MultRegCMP-package}
\alias{MultRegCMP}
\alias{MultRegCMP-package}
\title{MultRegCMP: Bayesian Multivariate Conway-Maxwell-Poisson Regression Model for Correlated Count Data}
\description{
Fits a Bayesian Regression Model for multivariate count data. This model assumes that the data is distributed according to the Conway-Maxwell-Poisson distribution (Conway and Maxwell 1962), and for each response variable it is associate different covariates. This model allows to account for correlations between the counts by using latent effects based on the Chib and Winkelmann (2001) proposal.
}
\keyword{internal}
