% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultRegCMP-package.R
\docType{package}
\name{MultRegCMP-package}
\alias{MultRegCMP}
\alias{MultRegCMP-package}
\title{MultRegCMP: MCMC estimation for a Multivariate Conway-Maxwell-Poisson Regression Model}
\description{
MCMC Algorithm to estimate the parameters in a multivariate Conway-Maxwell-Poisson model for correlated count data.
}
\keyword{internal}
