% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cancer_death_risk.R
\name{update_cancer_death_risk}
\alias{update_cancer_death_risk}
\title{Combine Risk Estimates}
\usage{
update_cancer_death_risk(risk_cancer_death, risk_other_cause, time)
}
\arguments{
\item{risk_cancer_death}{risk of death from cancer within time \verb{t=}}

\item{risk_other_cause}{risk of death from other causes within time \verb{t=}}

\item{time}{time frame associated with the risk predictions above.}
}
\value{
numeric probability between zero and one
}
\description{
Provided a risk estimate of death from cancer and an estimate for death
from other causes, this function will combine the two risks assuming
an exponential distribution for both.
}
\examples{
update_cancer_death_risk(0.5, 0.5, time = 10)
}
